Share API across host and player
