+++ code/simulator/extensions/dedicated-local-multiplayer/player/src/socket/socket_context.tsx
@@ -1,11 +1,10 @@
 import { createContext } from "react";
 import type { Socket } from "socket.io-client";
-// import { type PlayerToServerEvents, type ServerToPlayerEvents } from "../../../host/app";
-// TODO make this interface shared
+import { type PlayerToServerEvents, type ServerToPlayerEvents } from "../../../api";
 
 type SocketContextType = {
-  socket: Socket | null;
-  // socket: Socket<ServerToPlayerEvents, PlayerToServerEvents> | null;
+  // socket: Socket | null;
+  socket: Socket<ServerToPlayerEvents, PlayerToServerEvents> | null;
   setSocket: (value: Socket) => void;
 };
